pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energy-trading-ai"
version = "0.1.0"
description = "End-to-end energy trading AI system with deep learning, NLP, RAG, and algorithmic trading"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Energy Trading AI Contributors" }]
keywords = ["energy", "trading", "ai", "machine-learning", "nlp", "forecasting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    # Data ingestion
    "yfinance>=0.2.36",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "fredapi>=0.5.1",
    "feedparser>=6.0.11",
    # Feature engineering
    "ta>=0.11.0",
    "scikit-learn>=1.4.0",
    # NLP & Sentiment
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "sentence-transformers>=3.0.0",
    # RAG / LLM
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "unstructured>=0.13.0",
    "openai>=1.30.0",
    # ML / DL models
    "xgboost>=2.0.0",
    "pytorch-forecasting>=1.0.0",
    # Backtesting
    "vectorbt>=0.26.0",
    "pyportfolioopt>=1.5.5",
    # Dashboard
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    # Utilities
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "pre-commit>=3.7.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "ipywidgets>=8.1.0",
]

[project.scripts]
ingest-data = "scripts.ingest_data:main"
build-features = "scripts.build_features:main"
build-vector-store = "scripts.build_vector_store:main"
train-models = "scripts.train_models:main"
run-backtest = "scripts.run_backtest:main"
generate-report = "scripts.generate_report:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008", "N806", "N803"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"notebooks/*" = ["E402", "F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
no_strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -v"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "notebooks/*", "scripts/*"]
